% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{descargar_Archivo}
\alias{descargar_Archivo}
\title{Descargar Archivo}
\usage{
descargar_Archivo(tipo_eleccion, año, mes, ambito, directorio = "./descargas/")
}
\arguments{
\item{tipo_eleccion}{Tipo de elección. Puede ser 'referendum', 'congreso', 'senado', 'municipales', 'cabildos' o 'europeas'.}

\item{año}{Año de la elección.}

\item{mes}{Mes de la elección.}

\item{ambito}{Ámbito de la elección. Puede ser 'mesa', 'municipio' o 'superior'.}

\item{directorio}{Directorio donde se guardará el archivo descargado. Por defecto es "./descargas/".}
}
\value{
La ruta del directorio donde se ha descargado el archivo.
}
\description{
Descarga un archivo desde la fuente de datos electorales del gobierno
}
\details{
Esta función descarga un archivo en base a unos parametros que define cada elección, de los que se genera una URL de la que obtener los datos.
}
\examples{
# Descargar archivo de elecciones municipales del año 2023, mes 05, en el ámbito mesa
descargar_Archivo(tipo_eleccion = "municipales", año = 2023, mes = '05', ambito = "mesa")

}
\seealso{
https://infoelectoral.interior.gob.es/es/elecciones-celebradas/area-de-descargas/ URL base de descarga de archivos
}
\author{
Ignacio Domínguez Calvo
}
\keyword{archivo}
\keyword{descarga}
\keyword{elecciones}
\keyword{elección}
